chore: tweaked code  according to the code review comments
