+++ doc/zh-cn/plugins/server-info.md
@@ -25,13 +25,13 @@
 - [插件属性](#插件属性)
 - [**插件接口**](#插件接口)
 - [启用插件](#启用插件)
+- [如何自定义服务信息上报间隔](#如何自定义服务信息上报间隔)
 - [测试插件](#测试插件)
 - [禁用插件](#禁用插件)
-- [注意事项](#注意事项)
 
 ## 插件简介
 
-`server-info` 是 `APISIX` 提供的一款查询其服务基本信息的插件。
+`server-info` 是一款能够定期将服务基本信息上报至 etcd，同时允许我们通过它提供的 API 在数据面访问到这些数据的插件。
 
 ## 插件属性
 
@@ -56,7 +56,24 @@ plugins:                          # plugin list
   ......
 ```
 
-在启动/重启 APISIX 之后，即可通过访问 `/apisix/server_info` 来获取服务基本信息。
+在启动 APISIX 之后，即可通过访问 `/apisix/server_info` 来获取服务基本信息。
+
+## 如何自定义服务信息上报间隔
+
+我们可以在 `conf/config.yaml` 文件的 `plugin_attr` 一节中修改上报间隔。
+
+| 名称         | 类型   | 默认值  | 描述                                                          |
+| ------------ | ------ | -------- | -------------------------------------------------------------------- |
+| report_interval | number | 60 | 上报服务信息至 etcd 的间隔（单位：秒）|
+
+下面的例子将服务信息上报间隔修改成了 10 秒：
+
+```yaml
+plugin_attr:
+  server-info:
+    report_interval: 10
+```
+
 
 ## 测试插件
 
@@ -97,8 +114,3 @@ plugins:                          # plugin list
   - zipkin
   ......
 ```
-
-## 注意事项
-
-当使用 etcd 作为 APISIX 的数据中心的说话，服务信息将被周期性地上报到 etcd（目前的上报间隔是 5
-秒）并被 APISIX Dashboard 所收集，所以你也可以通过 APISIX Dashboard 来访问这些数据。
